splitting into small components
