--- src/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./dataimport";
-export * from "./header";
-export * from "./table";
-export * from "./elemenet";
