+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./version";
 export * from "./WidgetModel";
-export * from "./components/table";
-export * from "./components/dataimport"
+export * from "./table/table";
+export * from "./input/dataimport"
