+++ src/input/index.ts
@@ -1,0 +1,3 @@
+export * from "./dataimport";
+export * from "./outname";
+export * from "./inputarea";
